Fixed DN1 a)

+++ Homework/HW1/naloga1/DN01a_63200342.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdbool.h>
 
-void izpisi(int st)
+void izpisi(long st)
 {
     if (st < 0)
     {
@@ -17,7 +17,7 @@ int main(int argc, char const *argv[])
 {
 
     int c = getchar();
-    int val = 0;
+    long val = 0;
     int isNeg = 1;
     if (c == '-')
         isNeg = -1;
@@ -38,7 +38,7 @@ int main(int argc, char const *argv[])
     val *= isNeg;
 
     isNeg = 1;
-    int val2 = 0;
+    long val2 = 0;
     c = getchar();
     if (c == '-')
         isNeg = -1;
